Fixed visualization of IO-STG. Extended visualization with displaying input values of constraints. Fixed and elimination of SelfInInEdges. Added option --semantic-elimination for eliminating in-in self edges.

+++ projects/CodeThorn/src/Analyzer.h
@@ -201,7 +201,7 @@ namespace CodeThorn {
     map<string,VariableId> globalVarName2VarIdMapping;
     vector<bool> binaryBindingAssert;
     void setAnalyzerMode(AnalyzerMode am) { _analyzerMode=am; }
-    void setMaxTransitions(long int maxTransitions) { _maxTransitions=maxTransitions; }
+    void setMaxTransitions(size_t maxTransitions) { _maxTransitions=maxTransitions; }
     bool isIncompleteSTGReady();
   private:
     set<int> _inputVarValues;
@@ -221,7 +221,7 @@ namespace CodeThorn {
     int _solver;
 	AnalyzerMode _analyzerMode;
 	set<const EState*> _newNodesToFold;
-    long int _maxTransitions;
+    size_t _maxTransitions;
   };
   
 } // end of namespace CodeThorn
